// FE.Core.FW/src/apis/models/ExamScheduleTopikModel.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ExamScheduleTopikModel = {
  id?: string;
  examinationName?: string | null;
  examDate?: string | null;
  examDateString?: string | null;
  examTime?: string | null;
  examId?: string;
  examWorkShiftId?: string;
  startRegister?: string;
  endRegister?: string;
  note?: string | null;
  status?: number;
  public?: boolean;
  englishName?: string | null;
  koreaName?: string | null;
  examPeriodId?: string;
  noteTimeEnterExamRoom?: string;
};