// FE.Core.FW/src/apis/models/ManageApplicationTimeModel.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ManageApplicationTimeModel = {
  id?: string;
  sysTimeFrameId?: string;
  headerQuarterId?: string;
  maxRegistry?: number;
  receivedDate?: string;
  registed?: number;
  timeStart?: string | null;
  timeEnd?: string | null;
  isShow?: boolean;
  listReceivedDate?: Array<string> | null;
};